pyproject.toml
[project]
name = "dtc-capstone-music-streaming-app"
version = "0.1.0"
description = "Real-time music streaming analytics — Kafka, Flink, Iceberg, Trino, ClickHouse on GKE"
readme = "README.md"
requires-python = ">=3.11,<3.13"  # dbt + Airflow max support is 3.12 as of 2025
dependencies = [
  # --- dbt ---
  "dbt-core>=1.8",
  "dbt-bigquery>=1.8",                              # medallion silver/gold → BigQuery + Looker Studio
  "dbt-trino>=1.8",                                 # reserved for future Trino query layer
  # --- Apache Airflow (orchestration) ---
  "apache-airflow>=2.9",
  "apache-airflow-providers-trino>=5.0",
  "apache-airflow-providers-google>=10.0",        # GCS operators
  "apache-airflow-providers-apache-kafka>=1.3",   # Kafka sensors / operators
  "apache-airflow-providers-cncf-kubernetes>=8.0", # KubernetesPodOperator for Flink jobs
  # --- Kafka (Python producer for testing / eventsim health checks) ---
  "confluent-kafka>=2.4",
  # --- Apache Iceberg (Python catalog management + table inspection) ---
  "pyiceberg[gcs]>=0.7",
  # --- GCP ---
  "google-cloud-storage>=2.16",
  # --- Data quality ---
  "soda-core-trino>=3.3",                         # data quality checks in Airflow DAGs
  "diagrams>=0.25.1",
]

[dependency-groups]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.4",
  "pre-commit>=3.7",
  "ipython>=8.0",
]
